requirements.txt
python-dotenv>=1.0.1,<2.0.0
transformers
datasets
huggingface_hub[hf_transfer]
vllm
tenacity>=9.0.0
openai>=1.30.0
prettytable>=3.0.0
openpyxl>=3.1.5
anthropic>=0.47.2
jsonpickle>=4.0.2
pandas
timeout_decorator
pyarrow
pyyaml
easydict
smart_open[s3]
tiktoken
orjsonl
statsmodels
krippendorff
tos
tosfs
fsspec
